Updates after module rename
