+++ Solana/bridge-authority/integration-tests/src/test/kotlin/net/corda/samples/solana/bridge/authority/TestBase.kt
@@ -1,4 +1,4 @@
-package net.corda.samples.solana.bridge.token
+package net.corda.samples.solana.bridge.authority
 
 import com.r3.corda.lib.solana.core.AccountManagement
 import com.r3.corda.lib.solana.core.FileSigner
